+++ test/unit/test_interfaces.py
@@ -6,14 +6,14 @@ class TestInterfaces:
 
     @pytest.mark.sad
     def test_step(self):
-        from pipelayer.step.protocol import Step
+        from pipelayer.protocol.step import Step
 
         with pytest.raises(NotImplementedError):
             Step.run(None, None, None)
 
     @pytest.mark.sad
     def test_compound_step(self):
-        from pipelayer.compound_step.protocol import CompoundStep
+        from pipelayer.protocol.compound_step import CompoundStep
 
         with pytest.raises(NotImplementedError):
             CompoundStep._run_steps(None, None, None)